feat(catalog): add component demo suite and update package namespace
Update the catalog module's package namespace from framework.cortena.ui.catalog to app.cortena.ui.catalog in build configuration files and the project's architecture documentation.
Add the main catalog Activity to aggregate all component demos, plus individual demo screens for ScrollView, Toggle, Slider, Button, typography, and color system tokens.
Remove the legacy color demo file stored in the old package path.
